+++ src/test/scala/org/scalacheck/examples/MathSpec.scala
@@ -5,7 +5,7 @@ import org.scalacheck.Prop.{forAll, BooleanOperators}
 object MathSpec extends org.scalacheck.Properties("Math") {
   property("sqrt") = forAll { n: Int =>
     (n >= 0) ==> {
-      val m = math.sqrt(n)
+      val m = math.sqrt(n.toDouble)
       math.round(m*m) == n
     }   
   }
